refactor(sync): one fence over the same EventApi
`Fence` needs five driver calls and `EventApi` already had four of them; only
`stream_wait_event` is new. So the second copy of the fence goes the way the
second copy of the profiler did, and the trait now has the two callers that
justify it rather than one.

The duplication here was semantic rather than textual — the two files were
written independently and agreed on the five calls while disagreeing on
everything around them — so it also reaches past the two files: the identical
`impl drop_queue::Fence` in both `stream.rs` moves into `cubecl-runtime` with
the type.

Where the two disagreed, HIP's version won: the `expect` that says why a
fence which never recorded is fatal, and the module doc explaining that the
server's lock is the reason a fence exists at all. `unsafe impl Send for
Fence` is gone from both — `EventApi::Event: Send + Sync` already carries it.

One behavior change: releasing the event is now the shared `Event`'s `Drop`,
which logs. CUDA panicked on a failed `cuEventDestroy` in `wait_async` and
both backends reported one from `wait_sync` as a stream fault. Neither reads
right — a destroy that fails after the wait succeeded says nothing about the
work — and the profiler's events already logged.

`timings.rs` becomes `events.rs` in both backends: with the fence there too,
the file is the backend's event API and no longer only about timing.
