--- crates/cubecl-cuda/src/compute/sync/mod.rs
@@ -1,3 +1,0 @@
-mod fence;
-
-pub use fence::*;
