+++ crates/cubecl-hip/src/compute/events.rs
@@ -1,13 +1,18 @@
-//! HIP's half of device profiling: the driver calls, and the type naming them.
+//! HIP's device events: the driver calls, and the type naming them.
 //!
-//! The design — what a profiling window is, why its events are read back
-//! lazily, and how the anchor places one on the host clock — lives with the
-//! shared [`EventProfiler`](cubecl_runtime::device_events::EventProfiler).
+//! What is built on them — a [`Fence`] handed out so a caller can wait on a
+//! stream outside the server's lock, and an [`EventProfiler`] timing work on
+//! the device's own clock — lives with the shared
+//! [`device_events`](cubecl_runtime::device_events) module, along with the
+//! design arguments for both.
 
 use cubecl_common::profile::Duration;
 use cubecl_hip_sys::{hipEvent_t, hipStream_t};
 use cubecl_runtime::device_events::EventApi;
 use cubecl_runtime::driver::{DriverError, checked};
+
+/// A fence, over HIP's event API.
+pub type Fence = cubecl_runtime::device_events::EventFence<Hip>;
 
 /// The device profiler, over HIP's event API.
 pub type EventProfiler = cubecl_runtime::device_events::EventProfiler<Hip>;
@@ -22,9 +27,8 @@ pub struct Hip;
 pub struct HipEvent(hipEvent_t);
 
 // SAFETY: a `hipEvent_t` is a handle into the driver rather than thread-affine
-// state. It is recorded on the server's thread and read back on whichever
-// thread awaits the profile — the same crossing [`Fence`](super::fence::Fence)
-// already makes.
+// state. It is recorded on the server's thread and waited for on whichever
+// thread holds the fence or awaits the profile.
 unsafe impl Send for HipEvent {}
 unsafe impl Sync for HipEvent {}
