+++ crates/cubecl-hip/src/compute/mod.rs
@@ -3,10 +3,9 @@ mod server;
 pub(crate) mod capture;
 pub(crate) mod context;
 pub(crate) mod driver;
-pub(crate) mod fence;
+pub(crate) mod events;
 pub(crate) mod storage;
 pub(crate) mod stream;
-pub(crate) mod timings;
 
 pub use server::*;
 pub use storage::*;
